requirements.txt
flask
web3
python-dotenv
scikit-learn 
joblib
statsmodels
pandas
pandas_ta
numba
tqdm